fix(search): finish mainline qualification
