+++ crates/ctx-history-read-application/tests/contracts/search_show.rs
@@ -277,7 +277,6 @@ fn codex_ctx_retrieval_echoes_are_hidden_from_search_but_remain_directly_retriev
 fn search_keeps_huge_matching_grapheme_hits_in_json_and_human_output() {
     const SNIPPET_MAX_BYTES: usize = 16 * 1024;
     const SESSION_ID: &str = "019c0000-0000-7000-8000-000000000001";
-    const EXPECTED_FALLBACK: &str = "…/workspace/huge-grapheme";
 
     let temp = tempdir();
     let oversized_cluster = format!("x{}", "\u{301}".repeat(SNIPPET_MAX_BYTES));
@@ -319,16 +318,10 @@ fn search_keeps_huge_matching_grapheme_hits_in_json_and_human_output() {
         Some(1),
         "{first:#}"
     );
-    assert_eq!(first_result["snippet"], EXPECTED_FALLBACK, "{first:#}");
+    assert_eq!(first_result["snippet"], "", "{first:#}");
     assert_eq!(first_result["snippet_truncated"], true, "{first:#}");
     assert!(
-        !first_result["snippet"].as_str().unwrap().is_empty()
-            && first_result["snippet"].as_str().unwrap().chars().count() <= 320
-            && !first_result["snippet"]
-                .as_str()
-                .unwrap()
-                .contains('\u{301}')
-            && first_result["snippet"].as_str().unwrap().len() <= SNIPPET_MAX_BYTES,
+        first_result["snippet"].as_str().unwrap().len() <= SNIPPET_MAX_BYTES,
         "{first:#}"
     );
     assert_eq!(first_result["snippet"], second_result["snippet"]);
@@ -360,7 +353,7 @@ fn search_keeps_huge_matching_grapheme_hits_in_json_and_human_output() {
     let second_human = render_human();
     assert_eq!(first_human, second_human);
     assert!(first_human.contains("1 result"), "{first_human}");
-    assert!(first_human.contains(EXPECTED_FALLBACK), "{first_human}");
+    assert!(first_human.contains("codex user message"), "{first_human}");
     assert!(!first_human.contains('\u{301}'), "{first_human}");
 }
